requirements.txt
xai-sdk>=1.4.0
python-dotenv>=1.0.0

